requirements.txt
numpy
pandas
Pillow
torch
torchvision
